# 2015-2016-api-mobiliteitscentrum/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mobiliteitscentrum Gent Website API (VOPRO 2015-1016)",
        "description": "API specificatie voor Milestone 3.\n",
        "version": "3.0.0"
    },
    "host": "voproX.ugent.be",
    "schemes": [
        "http"
    ],
    "basePath": "/api",
    "produces": [
        "application/json"
    ],
    "paths": {
        "/user/": {
            "get": {
                "summary": "Returns all users",
                "description": "This endpoint returns all users in JSON format. Only accessible for Administrators (later use).\n",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The access token as a bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array of users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "The current user is not an Administrator",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new user",
                "description": "Creates a new user.\n",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User input parameters",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "A bearer to perform captcha validation",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The user is created",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "409": {
                        "description": "The user already exists (ex. email already in use)",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Unique identifier representing a specific User. This attribute is not required in a request, but is required in a response."
                },
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Link"
                    }
                },
                "password": {
                    "type": "string",
                    "description": "User password in clear text. Please save this in a secure way. Clearly, this attribute may not be returned in a response.\n"
                },
                "email": {
                    "type": "string",
                    "description": "Email address of the user."
                },
                "cell_number": {
                    "type": "string",
                    "description": "Cellphone number of the user."
                },
                "mute_notifications": {
                    "type": "boolean",
                    "description": "Master switch to disable all notifications (ex. when on holiday)"
                },
                "validated": {
                    "type": "object",
                    "description": "Indicates which fields the user has validated through external services",
                    "properties": {
                        "email": {
                            "type": "boolean",
                            "description": "Indicates if this user has validated his/her email"
                        },
                        "cell_number": {
                            "type": "boolean",
                            "description": "Indicates if this user has validated his/her cell number"
                        }
                    }
                }
            },
            "required": [
                "password",
                "email",
                "mute_notifications"
            ]
        },
...
        "Link": {
            "type": "object",
            "properties": {
                "rel": {
                    "type": "string",
                    "description": "Indicates the relation of the linked object (ex. self)"
                },
                "href": {
                    "type": "string",
                    "description": "The actual link in full URL (ex. https://vopro.ugent.be/api/user/123/)"
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32",
                    "description": "HTTP status code (ex. 400)"
                },
                "message": {
                    "type": "string",
                    "description": "Human readable error (ex. You are not logged in)"
                },
                "fields": {
                    "type": "string",
                    "description": "Missing or incorrect fields for input"
                }
            }
        },
...
    }
}